+++ src/core/examples/log-example.cc
@@ -51,9 +51,9 @@ FreeEvent()
 
     NS_LOG_ERROR("FreeEvent: error msg");
     NS_LOG_WARN("FreeEvent: warning msg");
-    NS_LOG_DEBUG("FreeEvent: debug msg");
     NS_LOG_INFO("FreeEvent: info msg");
     NS_LOG_LOGIC("FreeEvent: logic msg");
+    NS_LOG_DEBUG("FreeEvent: debug msg");
 }
 
 /** Simple object to aggregate to a node.
@@ -93,9 +93,9 @@ class MyEventObject : public Object
 
         NS_LOG_ERROR("MyEventObject:Event: error msg");
         NS_LOG_WARN("MyEventObject:Event: warning msg");
-        NS_LOG_DEBUG("MyEventObject:Event: debug msg");
         NS_LOG_INFO("MyEventObject:Event: info msg");
         NS_LOG_LOGIC("MyEventObject:Event: logic msg");
+        NS_LOG_DEBUG("MyEventObject:Event: debug msg");
     }
 
 }; // MyEventObject
@@ -110,26 +110,26 @@ main(int argc, char** argv)
     CommandLine cmd;
     cmd.Parse(argc, argv);
 
-    NS_LOG_UNCOND("Creating a Node");
+    NS_LOG_DEBUG("Creating a Node");
     auto node = CreateObject<Node>();
 
-    NS_LOG_UNCOND("Creating MyEventObject");
+    NS_LOG_DEBUG("Creating MyEventObject");
     auto myObj = CreateObject<MyEventObject>();
 
-    NS_LOG_UNCOND("Aggregating MyEventObject to Node");
+    NS_LOG_DEBUG("Aggregating MyEventObject to Node");
     node->AggregateObject(myObj);
 
-    NS_LOG_UNCOND("Scheduling the MyEventObject::Event with node context");
+    NS_LOG_INFO("Scheduling the MyEventObject::Event with node context");
     Simulator::ScheduleWithContext(node->GetId(), Seconds(3), &MyEventObject::Event, &(*myObj));
 
-    NS_LOG_UNCOND("Scheduling FreeEvent");
+    NS_LOG_INFO("Scheduling FreeEvent");
     Simulator::Schedule(Seconds(5), FreeEvent);
 
-    NS_LOG_UNCOND("Starting run...");
+    NS_LOG_DEBUG("Starting run...");
     Simulator::Run();
-    NS_LOG_UNCOND("... run complete");
+    NS_LOG_DEBUG("... run complete");
     Simulator::Destroy();
-    NS_LOG_UNCOND("Goodbye");
+    NS_LOG_DEBUG("Goodbye");
 
     return 0;
 }